requirements.txt
# ═══════════════════════════════════════════════════════════════
# Chat-with-PDF Advanced AI — Dependencies
# Python >= 3.10 required
# ═══════════════════════════════════════════════════════════════

# ── Core Framework ────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ── LangChain Ecosystem ───────────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ── OpenAI ────────────────────────────────────────────────────
openai>=1.30.0

# ── Vector Stores ─────────────────────────────────────────────
faiss-cpu>=1.8.0
chromadb>=0.5.0

# ── PDF Processing ────────────────────────────────────────────
pymupdf>=1.24.0          # PyMuPDF — primary PDF parser
pdfplumber>=0.11.0       # Fallback PDF parser
pypdf>=4.2.0             # Additional PDF utilities
pytesseract>=0.3.10      # OCR for scanned PDFs
Pillow>=10.3.0           # Image processing for OCR

# ── Embeddings & ML ───────────────────────────────────────────
sentence-transformers>=3.0.0   # Cross-encoder reranking
scikit-learn>=1.5.0            # BM25 utilities, metrics
numpy>=1.26.0
scipy>=1.13.0

# ── BM25 Sparse Retrieval ─────────────────────────────────────
rank-bm25>=0.2.2

# ── Text Processing ───────────────────────────────────────────
tiktoken>=0.7.0          # Token counting
nltk>=3.8.1              # NLP utilities
spacy>=3.7.0             # NLP pipeline
regex>=2024.5.15

# ── Data & Visualization ──────────────────────────────────────
pandas>=2.2.0
plotly>=5.22.0
altair>=5.3.0

# ── Async & Performance ───────────────────────────────────────
httpx>=0.27.0
aiohttp>=3.9.0
tenacity>=8.3.0          # Retry logic with exponential backoff
cachetools>=5.3.0        # In-memory caching

# ── Storage & Serialization ───────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# ── Dev Tools ─────────────────────────────────────────────────
black>=24.4.0
isort>=5.13.0
ruff>=0.4.0
